requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-dotenv==1.0.0
google-api-python-client==2.97.0
lxml==4.9.3
feedparser==6.0.10
